requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pycryptodome==3.19.0
python-chess==1.999